Fix a race condition
When the seats were being added on mobile, all the tokens would be in
the top-left corner. This was because the dimensions of the grimoire pad
was 0x0 (because it was hidden). The tokens couldn't be positioned using
Vue's reactivity because dragging them using the mouse should break it.
This commit resolves the various race conditions that will cause that.